notifications/django_admin/requirements/base.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-model-utils==4.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.4
psycopg2-binary==2.9.3
python-dotenv==0.15.0
python-logstash==0.4.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
sentry-sdk[django]==1.3.0

